+++ Contents/Code/common_upd.py
@@ -3,8 +3,9 @@
 import re, datetime
 import requests             # [а к нему еще chardet, urllib3, certifi, idna]      # type: ignore 
 
-from config import API_BASE_URL, FILM_DETAILS, FILM_DISTRIBUTION, FILM_POSTERS, FILM_REVIEW, FILM_STAFF, SERIAL_SEASONS
-from debug import d, w, log_timing, inspect_obj
+from config import API_BASE_URL, FILM_DETAILS, FILM_DISTRIBUTION, FILM_POSTERS, FILM_REVIEW, FILM_STAFF
+from config import SERIAL_SEASONS, UNKNOWN_YEAR, MAX_POSTERS_QTY
+from debug import d, w, log_timing       #  inspect_obj
 from utils import get_json, getGUIDs
 
 
@@ -18,7 +19,6 @@ def load_metadata(metadata, media, valid_names):
   movie_data = get_json(API_BASE_URL + FILM_DETAILS % metadata.id)
   if not movie_data or not metadata.id:
     return
-  
   # добавить   "imdbId": "tt0379786",
   # now media.guid like :kp722886:
   Guids = getGUIDs(media.guid)
@@ -36,70 +36,57 @@ def load_metadata(metadata, media, valid_names):
 
   #   title                     = Template.String()
   repls = (u' (видео)', u' (ТВ)', u' (мини-сериал)', u' (сериал)')
-  title = reduce(lambda a, kv: a.replace(kv, ''), repls, movie_data.get('nameRu')) # type: ignore
+  title = reduce(lambda a, kv: a.replace(kv, ''), repls, movie_data.get('nameRu', '')) # type: ignore
   metadata.title = title
 
   if isAgentMovies:
     #   year                      = Template.Integer()
-    if movie_data.get('year'):
-      metadata.year = int(movie_data.get('year'))
-    else:
-      metadata.year = ''
-    if metadata.year == 0:
-      metadata.year = ''
+    metadata.year = movie_data.get('year', str(UNKNOWN_YEAR))
     
     #   originally_available_at   = Template.Date()   A dateobject specifying the movie’s original release date.
     #   tagline                   = Template.String()
-    metadata.tagline = movie_data.get('slogan')
-    
+    metadata.tagline = movie_data.get('slogan', '')
   #   summary                   = Template.String()
   summary_add = ''
   # "Описание: отображать слоган фильма"
   if Prefs['desc_show_slogan']: # type: ignore
-    val = movie_data.get('shortDescription')
+    val = movie_data.get('shortDescription', '')
     if val:
       summary_add += "%s\n" % val
-      d(u"   Девиз = %s" % summary_add)
+      d(u"desc_show_slogan:   Девиз = %s" % summary_add)
   #  "Описание: отображать рейтинг Кинопоиск"
   if Prefs['desc_rating_kp']: # type: ignore
-    val = movie_data.get('ratingKinopoisk')
+    val = movie_data.get('ratingKinopoisk', '')
     if val:
       summary_add += u'КиноПоиск: %s' % val
-      val1 = movie_data.get('ratingKinopoiskVoteCount')
+      val1 = movie_data.get('ratingKinopoiskVoteCount', '')
       if val1:
         summary_add += ' (%s)' % val1
-      d(u"   'КиноПоиск: %s(%s) " % (val, val1))
+      d(u"desc_rating_kp:   'КиноПоиск: %s(%s) " % (val, val1))
       summary_add += '\n'
   # "Описание: отображать рейтинг IMDB"
   if Prefs['desc_rating_imdb']: # type: ignore
-    val = movie_data.get('ratingImdb')
+    val = movie_data.get('ratingImdb', '')
     if val:
       summary_add += u'IMDB: %s' % val
-      val1 = movie_data.get('ratingImdbVoteCount')
+      val1 = movie_data.get('ratingImdbVoteCount', '')
       if val1:
         summary_add += ' (%s)' % val1
       summary_add += '\n'
   #
-  try:
-    summary_add += movie_data.get('description')
-  except:  pass
+  summary_add += movie_data.get('description', '')
   #[Feature request] ссылка на инфо о фильме на кинопоиск #32
-  if Prefs['desc_add_links']: # type: ignore
-    summary_add += "\n" + movie_data.get('webUrl') + "\n"
-    if movie_data.get('imdbId'):
+  if Prefs['desc_add_links']:       # type: ignore
+    summary_add += "\n" + movie_data.get('webUrl', '') + "\n"
+    if movie_data.get('imdbId', ''):
       summary_add += 'https://www.imdb.com/title/' + movie_data.get('imdbId') + '/'
   # Всё, что сложилось - в описание
   metadata.summary = summary_add
   
   ## rating     A float between 0 and 10 specifying the movie’s rating.
-  metadata.rating = 0.0
-  try:
-    metadata.rating = float(movie_data.get('ratingImdb'))
-  except:  pass
+  metadata.rating = float(movie_data.get('ratingImdb', '') or 0)
   if metadata.rating == 0:
-    try:
-      metadata.rating = float('ratingKinopoisk')
-    except:  pass 
+    metadata.rating = float(movie_data.get('ratingKinopoisk', '') or 0)
   '''
   "Rating": [
           {
@@ -125,52 +112,45 @@ def load_metadata(metadata, media, valid_names):
         ]
   '''
 
-
   #   trivia                    = Template.String()   A string containing trivia about the movie.
   #   quotes                    = Template.String()
   #   content_rating            = Template.String()
-  val = movie_data.get('ratingMpaa')
-  if val:
-    metadata.content_rating = val.upper()
+  metadata.content_rating = movie_data.get('ratingMpaa', '')
+  if metadata.content_rating:
+    metadata.content_rating = metadata.content_rating.upper()
   else:
-    val = movie_data.get('ratingAgeLimits')
-    metadata.content_rating = val
-  #   content_rating_age        = Template.Integer()
-  try:
-    metadata.content_rating_age = int(movie_data.get('ratingAgeLimits').replace('age', '') or 0)
-  except:  pass
+    metadata.content_rating = movie_data.get('ratingAgeLimits', '')
+  # 'TV_Show' object has no attribute named 'content_rating_age'
+  if isAgentMovies:
+    val = movie_data.get('ratingAgeLimits', 0)
+    if val:
+      metadata.content_rating_age = int(val.replace('age', ''))
+
   #   countries                 = Template.Set(Template.String())
-  countries = []
-  for country in movie_data.get('countries'):
-    countries.append(country.get('country'))
-  metadata.countries = countries
+  if 'country' in movie_data:
+    metadata.countries =[]
+    for country in movie_data.get('countries'):
+      metadata.countries.append(country.get('country'), '')
   
   #   chapters                  = Template.Set(Chapter()) 
   ##  original_title       A string specifying the movie’s original title.
-  val = movie_data.get('nameOriginal') 
-  metadata.original_title = val if val else movie_data.get('nameEn')
+  val = movie_data.get('nameOriginal', '') 
+  metadata.original_title = val if val else movie_data.get('nameEn', '')
   
   ## genres        A set of strings specifying the movie’s genre  
-  genres = []
-  for genre in movie_data.get('genres'):
-      genres.append(genre.get('genre'))
-  metadata.genres = genres       
+  arr = []
+  for genre in movie_data.get('genres', []):
+    arr.append(genre.get('genre', ''))
+  metadata.genres = arr
   
   ## posters     A container of proxy objects representing the movie’s posters. See below for information about proxy objects.
-  
-  try:
-    d("   === load meta start posterUrlPreview %s" % movie_data.get('posterUrlPreview'))
-    #c "c:\Users\plugo\AppData\Local\Plex Media Server\Plug-ins\MoviePosterDB.bundle\Contents\Code\__init__.py" 
-    image_url = movie_data.get('posterUrl')
-    thumb_url = movie_data.get('posterUrlPreview')
-    #metadata.posters[image_url] = Proxy.Preview(HTTP.Request(thumb_url), sort_order=0)
+  d("   === load meta start posterUrlPreview %s" % movie_data.get('posterUrlPreview'))
+  image_url = movie_data.get('posterUrl', '')
+  thumb_url = movie_data.get('posterUrlPreview', '')
+  if image_url and thumb_url:
     metadata.posters[image_url] = Proxy.Preview(requests.get(thumb_url).content, sort_order=0) # type: ignore
     valid_names.append(image_url)
-    #metadata.posters.validate_keys(valid_names)  #только после ВСЕХ posters
-  except:    pass
-  # d("===================update:load_meta end, Duration=%s\n"  % (getMilliseconds(Datetime.Now()) - msStart))
-  ## themes    A container of proxy objects representing the movie’s theme music. 
-
+    
 
 @log_timing
 def load_episodes(metadata, media):
@@ -187,33 +167,31 @@ def load_episodes(metadata, media):
     d(u"\nsrch_and_score: Попытка поиска без ключа? %s" % seasons_json["message"])
     return
   
-  
   #seasons_qty = seasons_json.get('total')
-  for season_num, season_data in enumerate(seasons_json.get('items'), 1):    # 1,2...
+  for season_num, season_data in enumerate(seasons_json.get('items', []), 1):    # 1,2...
     d("Try process season_num %s in json" % season_num)
     if season_num not in media.seasons:      
         continue
     # отлично, есть локальные файлы из сезонов из найденного сезона
     Log("есть локальные файлы из season_num %s" % season_num)      # type: ignore
-    for episode_num, episode_data in enumerate(season_data.get('episodes'), 1):
+    for episode_num, episode_data in enumerate(season_data.get('episodes', []), 1):
       # по каждой серии   =Season2set or 
-      s_num = episode_data.get('seasonNumber')
-      e_num = episode_data.get('episodeNumber')
+      s_num = episode_data.get('seasonNumber', '')
+      e_num = episode_data.get('episodeNumber', '')
+      if not (s_num and e_num):
+        w('OOPS! Эпизод %s сезона %s' % (e_num, s_num))
+        continue
       Log("Эпизод %s сезона %s" % (e_num, s_num))      # type: ignore
       episode = metadata.seasons[s_num].episodes[e_num]
       episode.title = ''
       episode.originally_available_at = None
       episode.summary = ''
       #
-      title = episode_data.get('nameRu') or ''
-      titleE = episode_data.get('nameEn') or ''
-      d("title:%s, titleE:%s" % (title, titleE))
-      if not title and titleE:
-        title = titleE
-      else:
-        if titleE:
-          title = title + ' / ' + titleE
+      title = episode_data.get('nameRu', '') or episode_data.get('nameEn', '')
+      if episode_data.get('nameRu', '') and episode_data.get('nameEn', ''):
+        title += ' / ' + episode_data.get('nameEn', '')
       episode.title = title
+      #
       dat = episode_data.get('releaseDate') or ''
       if dat:
         episode.originally_available_at = datetime.datetime.strptime(dat, "%Y-%m-%d").date()
@@ -231,11 +209,13 @@ def load_distribution(metadata):
     return
   #[ LOCAL, COUNTRY_SPECIFIC, PREMIERE, ALL, WORLD_PREMIER ], subType : [ CINEMA, DVD, DIGITAL, BLURAY ] 
   for data in movie_data:
-    if data.get('type') == 'ALL':
-      metadata.originally_available_at = Datetime.ParseDate((data.get('date')).replace('00.', '01.'), '%Y-%m-%d').date()  # type: ignore
+    if data.get('type', '') == 'ALL':
+      dat = data.get('date', '')
+      if dat:
+        metadata.originally_available_at = Datetime.ParseDate(dat.replace('00.', '01.'), '%Y-%m-%d').date()  # type: ignore
       #   studio                    = Template.String() A string specifying the movie’s studio.
-      for comp in data.get('companies'):
-        metadata.studio = comp.get('name')  # so, only last like highlander  1981-12-02  Studio updated
+      for comp in data.get('companies', []):
+        metadata.studio = comp.get('name', '')  # so, only last like highlander  1981-12-02  Studio updated
   # d("===================update:load_distribution end, Duration=%s\n"  % (getMilliseconds(Datetime.Now()) - msStart))
     
   
@@ -268,10 +248,10 @@ def load_staff(metadata):
       @task # type: ignore
       def upd_member(metadata = metadata,
                      member = member):
-        prof = member.get('professionKey')
+        prof = member.get('professionKey', '')
         name = member.get('nameRu') if member.get('nameRu') else member.get('nameEn')
         url = member.get('posterUrl')
-        role_description = member.get('description')
+        role_description = member.get('description', '')
         if prof == 'DIRECTOR' and isAgentMovies:
           if name:
             director = metadata.directors.new()
@@ -309,7 +289,7 @@ def load_reviews(metadata):
   
   # Описание: загружать отзывы
   if not Prefs['desc_load_reviews']: # type: ignore
-    d(u"Обзоры не загружать (настройка Описание: загружать отзывы = false)")
+    d(u"Отзывы не загружать (настройка Описание: загружать отзывы = false)")
     metadata.reviews.clear()
     return
 
@@ -319,7 +299,7 @@ def load_reviews(metadata):
   @parallelize  # type: ignore
   def upd_reviews():
     metadata.reviews.clear()
-    for item in reviews_dict.get('items'):
+    for item in reviews_dict.get('items', []):
       @task # type: ignore
       def add_review(item = item,
                      metadata = metadata):
@@ -383,7 +363,7 @@ def load_gallery(metadata, valid_names, IsNeedPosters = True):
           art_dict = load_dict('STILL')
           if art_dict['total'] == 0:
             #тут рыбы нет, ну вообще нет
-            d("load_gallery: nothing %s finded")
+            w("load_gallery: nothing %s finded")
             return
 
   d("film='%s' id = %s, find total: %s" % (metadata.title, metadata.id, art_dict['total']))  
@@ -391,10 +371,10 @@ def load_gallery(metadata, valid_names, IsNeedPosters = True):
   @parallelize   # type: ignore
   def upd_posters():
     pref_max = int(Prefs['poster_limit'])         # type: ignore
-    max_posters = pref_max if pref_max else 30    # ну вот захардкодил 30 как 'очень много'
-    for i, mov in enumerate(art_dict.get('items')):
-      image_url = mov.get('imageUrl')
-      thumb_url = mov.get('previewUrl')
+    max_posters = pref_max if pref_max else MAX_POSTERS_QTY   # MAX_POSTERS_QTY как 'очень много'
+    for i, mov in enumerate(art_dict.get('items', [])):
+      image_url = mov.get('imageUrl', '')
+      thumb_url = mov.get('previewUrl', '')
       @task  # type: ignore # Create a task for updating poster
       def upd_poster( i=i, 
                       image_url=image_url,